tjpb/python_scripts/requirements.txt
astroid==2.3.3
camelot-py==0.7.3
chardet==3.0.4
click==7.1.1
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.3
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.3
pdfminer.six==20200402
pycryptodome==3.9.7
pylint==2.4.4
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sortedcontainers==2.1.0
typed-ast==1.4.1
wrapt==1.11.2
